+++ Map.cpp
@@ -1,0 +1,22 @@
+#include "Map.h"
+#include <iostream>
+
+
+void Map::Draw() {
+    for (int i = 0; i < _body.size(); i++) {
+        for (int j = 0; j < _body[i].size(); j++) {
+            switch (_body[i][j]) {
+                case ' ':
+                    std::cout << " ";  // Empty space
+                    break;
+                case 'R':
+                    std::cout << "R";  // Room
+                    break;
+                case 'C':
+                    std::cout << "C";  // Corridor
+                    break;
+            }
+        }
+        std::cout << std::endl;
+    }
+}
